mod to change default pcent check from int to float

+++ piranha/command.py
@@ -49,7 +49,7 @@ def main(sysargs = sys.argv[1:]):
     analysis_group.add_argument("-n","--min-read-length",action="store",type=int,help=f"Minimum read length. Default: {READ_LENGTH_DEFAULT_VP1[0]}")
     analysis_group.add_argument("-x","--max-read-length",action="store",type=int,help=f"Maximum read length. Default: {READ_LENGTH_DEFAULT_VP1[1]}")
     analysis_group.add_argument("-d","--min-read-depth",action="store",type=int,help=f"Minimum read depth required for consensus generation. Default: {VALUE_MIN_READS}")
-    analysis_group.add_argument("-p","--min-read-pcent",action="store",type=int,help=f"Minimum percentage of sample required for consensus generation. Default: {VALUE_MIN_PCENT}")
+    analysis_group.add_argument("-p","--min-read-pcent",action="store",type=float,help=f"Minimum percentage of sample required for consensus generation. Default: {VALUE_MIN_PCENT}")
     analysis_group.add_argument("--primer-length",action="store",type=int,help=f"Length of primer sequences to trim off start and end of reads. Default: {VALUE_PRIMER_LENGTH}")
 
     o_group = parser.add_argument_group('Output options')
